+++ cmd/retry/retry.go
@@ -1,14 +1,12 @@
 package cmd
 
 import (
-	"context"
 	"log"
 	"os"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/service/codebuild"
 	root "github.com/koh-sh/codebuild-multirunner/cmd"
-	mr "github.com/koh-sh/codebuild-multirunner/internal/multirunner"
+	cb "github.com/koh-sh/codebuild-multirunner/internal/codebuild"
 	"github.com/spf13/cobra"
 )
 
@@ -24,11 +22,11 @@ var retryCmd = &cobra.Command{
 	Use:   "retry",
 	Short: "retry CodeBuild build with a provided id",
 	Run: func(cmd *cobra.Command, args []string) {
-		client, err := mr.NewCodeBuildAPI()
+		client, err := cb.NewCodeBuildAPI()
 		if err != nil {
 			log.Fatal(err)
 		}
-		buildid, err := retryCodeBuild(client, id)
+		buildid, err := cb.RetryCodeBuild(client, id)
 		if err != nil {
 			log.Fatal(err)
 		}
@@ -45,7 +43,7 @@ var retryCmd = &cobra.Command{
 			}
 			time.Sleep(time.Duration(pollsec) * time.Second)
 			failed := false
-			ids, failed, err = mr.BuildStatusCheck(client, ids)
+			ids, failed, err = cb.BuildStatusCheck(client, ids)
 			if err != nil {
 				log.Fatal(err)
 			}
@@ -65,16 +63,4 @@ func init() {
 	retryCmd.Flags().IntVar(&pollsec, "polling-span", 60, "polling span in second for builds status check")
 	retryCmd.Flags().StringVar(&id, "id", "", "CodeBuild build id for retry")
 	retryCmd.MarkFlagRequired("id")
-}
-
-// retry CodeBuild build
-func retryCodeBuild(client mr.CodeBuildAPI, id string) (string, error) {
-	input := codebuild.RetryBuildInput{Id: &id}
-	result, err := client.RetryBuild(context.TODO(), &input)
-	if err != nil {
-		return "", err
-	}
-	buildid := *result.Build.Id
-	log.Printf("%s [STARTED]\n", buildid)
-	return buildid, err
 }
