+++ cmd/run/run.go
@@ -1,16 +1,13 @@
 package cmd
 
 import (
-	"context"
 	"log"
 	"os"
 	"time"
 
-	"github.com/aws/aws-sdk-go-v2/service/codebuild"
-	"github.com/jinzhu/copier"
 	root "github.com/koh-sh/codebuild-multirunner/cmd"
+	cb "github.com/koh-sh/codebuild-multirunner/internal/codebuild"
 	mr "github.com/koh-sh/codebuild-multirunner/internal/multirunner"
-	"github.com/koh-sh/codebuild-multirunner/internal/types"
 	"github.com/spf13/cobra"
 )
 
@@ -29,18 +26,18 @@ var runCmd = &cobra.Command{
 		if err != nil {
 			log.Fatal(err)
 		}
-		client, err := mr.NewCodeBuildAPI()
+		client, err := cb.NewCodeBuildAPI()
 		if err != nil {
 			log.Fatal(err)
 		}
 		ids := []string{}
 		hasfailedbuild := false
 		for _, v := range bc.Builds {
-			startbuildinput, err := convertBuildConfigToStartBuildInput(v)
+			startbuildinput, err := mr.ConvertBuildConfigToStartBuildInput(v)
 			if err != nil {
 				log.Fatal(err)
 			}
-			id, err := runCodeBuild(client, startbuildinput)
+			id, err := cb.RunCodeBuild(client, startbuildinput)
 			if err != nil {
 				log.Println(err)
 				hasfailedbuild = true
@@ -59,7 +56,7 @@ var runCmd = &cobra.Command{
 			}
 			time.Sleep(time.Duration(pollsec) * time.Second)
 			failed := false
-			ids, failed, err = mr.BuildStatusCheck(client, ids)
+			ids, failed, err = cb.BuildStatusCheck(client, ids)
 			if err != nil {
 				log.Fatal(err)
 			}
@@ -77,25 +74,4 @@ func init() {
 	root.RootCmd.AddCommand(runCmd)
 	runCmd.Flags().BoolVar(&nowait, "no-wait", false, "specify if you don't need to follow builds status")
 	runCmd.Flags().IntVar(&pollsec, "polling-span", 60, "polling span in second for builds status check")
-}
-
-// run CodeBuild Projects and return build id
-func runCodeBuild(client mr.CodeBuildAPI, input codebuild.StartBuildInput) (string, error) {
-	result, err := client.StartBuild(context.TODO(), &input)
-	if err != nil {
-		return "", err
-	}
-	id := *result.Build.Id
-	log.Printf("%s [STARTED]\n", id)
-	return id, nil
-}
-
-// copy configration read from yaml to codebuild.StartBuildInput
-func convertBuildConfigToStartBuildInput(build types.Build) (codebuild.StartBuildInput, error) {
-	startbuildinput := codebuild.StartBuildInput{}
-	err := copier.CopyWithOption(&startbuildinput, build, copier.Option{IgnoreEmpty: true, DeepCopy: true})
-	if err != nil {
-		return startbuildinput, err
-	}
-	return startbuildinput, nil
 }
