+++ internal/multirunner/multirunner_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/aws/aws-sdk-go-v2/service/codebuild/types"
+	"github.com/aws/aws-sdk-go-v2/service/codebuild"
 	"github.com/fatih/color"
 	cmt "github.com/koh-sh/codebuild-multirunner/internal/types"
 )
@@ -69,7 +69,7 @@ func Test_readConfigFile(t *testing.T) {
 	}
 }
 
-func Test_coloredString(t *testing.T) {
+func Test_ColoredString(t *testing.T) {
 	type args struct {
 		status string
 	}
@@ -101,78 +101,8 @@ func Test_coloredString(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := coloredString(tt.args.status); got != tt.want {
+			if got := ColoredString(tt.args.status); got != tt.want {
 				t.Errorf("coloredString() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_BuildStatusCheck(t *testing.T) {
-	id1 := "project:12345678"
-	id2 := "project2:87654321"
-	ids := []string{id1, id2}
-	errids := []string{"error"}
-
-	type args struct {
-		client func(t *testing.T) CodeBuildAPI
-		ids    []string
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    []string
-		want2   bool
-		wantErr bool
-	}{
-		{
-			name:    "all builds ended",
-			args:    args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "SUCCEEDED", Id: &id2}}), ids: ids},
-			want:    []string{},
-			want2:   false,
-			wantErr: false,
-		},
-		{
-			name:    "one builds in progress",
-			args:    args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "IN_PROGRESS", Id: &id2}}), ids: ids},
-			want:    []string{id2},
-			want2:   false,
-			wantErr: false,
-		},
-		{
-			name:    "one of builds failed",
-			args:    args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "FAILED", Id: &id2}}), ids: ids},
-			want:    []string{},
-			want2:   true,
-			wantErr: false,
-		},
-		{
-			name:    "one of builds timeout",
-			args:    args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "TIMED_OUT", Id: &id2}}), ids: ids},
-			want:    []string{},
-			want2:   true,
-			wantErr: false,
-		},
-		{
-			name:    "api error",
-			args:    args{client: ReturnBatchGetBuildsMockAPI([]types.Build{}), ids: errids},
-			want:    nil,
-			want2:   true,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got2, err := BuildStatusCheck(tt.args.client(t), tt.args.ids)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("buildStatusCheck() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got, tt.want)
-			}
-			if !reflect.DeepEqual(got2, tt.want2) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got2, tt.want2)
 			}
 		})
 	}
@@ -216,6 +146,37 @@ func Test_DumpConfig(t *testing.T) {
 			}
 			if got != tt.want {
 				t.Errorf("dumpConfig() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_ConvertBuildConfigToStartBuildInput(t *testing.T) {
+	type args struct {
+		build cmt.Build
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    codebuild.StartBuildInput
+		wantErr bool
+	}{
+		{
+			name:    "basic",
+			args:    args{cmt.Build{}},
+			want:    codebuild.StartBuildInput{},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ConvertBuildConfigToStartBuildInput(tt.args.build)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("convertBuildConfigToStartBuildInput() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("convertBuildConfigToStartBuildInput() = %v, want %v", got, tt.want)
 			}
 		})
 	}
